Move logic to the evaluator to avoid duplication
