+++ sqlmesh/core/plan/evaluator.py
@@ -21,7 +21,6 @@ from sqlmesh.core import analytics
 from sqlmesh.core import constants as c
 from sqlmesh.core.console import Console, get_console
 from sqlmesh.core.environment import EnvironmentNamingInfo
-from sqlmesh.core.model.definition import render_statements
 from sqlmesh.core.notification_target import (
     NotificationTarget,
 )
@@ -122,7 +121,16 @@ class BuiltInPlanEvaluator(PlanEvaluator):
                 deployability_index_for_evaluation = DeployabilityIndex.all_deployable()
 
             self.state_sync.update_project_statements(plan)
-            self._execute_project_statements(plan, snapshots_by_name, "before_all")
+            self.snapshot_evaluator._execute_project_statements(
+                plan.project_statements or [],
+                execution_stage="before_all",
+                snapshots=snapshots_by_name,
+                start=plan.start,
+                end=plan.end,
+                execution_time=plan.execution_time,
+                default_catalog=self.default_catalog,
+                environment_naming_info=plan.environment.naming_info,
+            )
 
             self._push(plan, snapshots, deployability_index_for_creation)
             update_intervals_for_new_snapshots(plan.new_snapshots, self.state_sync)
@@ -149,7 +157,16 @@ class BuiltInPlanEvaluator(PlanEvaluator):
             if not plan.requires_backfill:
                 self.console.log_success("Virtual Update executed successfully")
 
-            self._execute_project_statements(plan, snapshots_by_name, "after_all")
+            self.snapshot_evaluator._execute_project_statements(
+                plan.project_statements or [],
+                execution_stage="after_all",
+                snapshots=snapshots_by_name,
+                start=plan.start,
+                end=plan.end,
+                execution_time=plan.execution_time,
+                default_catalog=self.default_catalog,
+                environment_naming_info=plan.environment.naming_info,
+            )
 
         except Exception as e:
             analytics.collector.on_plan_apply_end(plan_id=plan.plan_id, error=e)
@@ -464,34 +481,6 @@ class BuiltInPlanEvaluator(PlanEvaluator):
                 )
 
         return snapshots_to_restate
-
-    def _execute_project_statements(
-        self, plan: EvaluatablePlan, snapshots: t.Dict[str, Snapshot], execution_stage: str
-    ) -> None:
-        adapter = self.snapshot_evaluator.adapter
-        if (statements := plan.project_statements) and (
-            rendered_expressions := [
-                expr
-                for project_statements in statements
-                for expr in render_statements(
-                    statements=getattr(project_statements, execution_stage),
-                    dialect=adapter.dialect,
-                    default_catalog=self.default_catalog,
-                    python_env=project_statements.python_env,
-                    start=plan.start,
-                    end=plan.end,
-                    execution_time=plan.execution_time or now(),
-                    snapshots=snapshots,
-                    environment_naming_info=plan.environment.naming_info,
-                )
-            ]
-        ):
-            with adapter.transaction():
-                for expr in rendered_expressions:
-                    adapter.execute(expr)
-            self.console.log_success(
-                f"Project's {execution_stage} statements executed successfully"
-            )
 
 
 class BaseAirflowPlanEvaluator(PlanEvaluator):
