+++ sqlmesh/core/snapshot/evaluator.py
@@ -40,6 +40,7 @@ from sqlmesh.core.audit import Audit, StandaloneAudit
 from sqlmesh.core.dialect import schema_
 from sqlmesh.core.engine_adapter import EngineAdapter
 from sqlmesh.core.engine_adapter.shared import InsertOverwriteStrategy
+from sqlmesh.core.loader import ProjectStatements
 from sqlmesh.core.macros import RuntimeStage
 from sqlmesh.core.model import (
     AuditResult,
@@ -51,6 +52,7 @@ from sqlmesh.core.model import (
     ViewKind,
     CustomKind,
 )
+from sqlmesh.core.model.definition import render_statements
 from sqlmesh.core.schema_diff import has_drop_alteration, get_dropped_column_names
 from sqlmesh.core.snapshot import (
     DeployabilityIndex,
@@ -1102,6 +1104,37 @@ class SnapshotEvaluator:
             physical_properties=rendered_physical_properties,
         )
         adapter.execute(snapshot.model.render_post_statements(**create_render_kwargs))
+
+    def _execute_project_statements(
+        self,
+        statements: t.List[ProjectStatements],
+        execution_stage: str,
+        snapshots: t.Optional[t.Dict[str, Snapshot]] = None,
+        execution_time: t.Optional[TimeLike] = None,
+        start: t.Optional[TimeLike] = None,
+        end: t.Optional[TimeLike] = None,
+        default_catalog: t.Optional[str] = None,
+        environment_naming_info: t.Optional[EnvironmentNamingInfo] = None,
+    ) -> None:
+        adapter = self.adapter
+        if rendered_expressions := [
+            expr
+            for project_statements in statements
+            for expr in render_statements(
+                statements=getattr(project_statements, execution_stage),
+                dialect=adapter.dialect,
+                default_catalog=default_catalog,
+                python_env=project_statements.python_env,
+                snapshots=snapshots,
+                start=start,
+                end=end,
+                execution_time=execution_time,
+                environment_naming_info=environment_naming_info,
+            )
+        ]:
+            with adapter.transaction():
+                for expr in rendered_expressions:
+                    adapter.execute(expr)
 
 
 def _evaluation_strategy(snapshot: SnapshotInfoLike, adapter: EngineAdapter) -> EvaluationStrategy:
